flask_app/requirements.txt
Flask==0.12.2
google-cloud-storage==1.2.0
gunicorn==19.7.1
tensorflow==1.2.1
numpy==1.13.0
scipy==0.19.1
Keras==2.0.2
h5py==2.6.0
Pillow==4.0.0